زompleting the initial version of the system

+++ SDSS_Laravel/database/migrations/2025_03_19_110805_create_configurations_table.php
@@ -17,8 +17,8 @@ return new class extends Migration
             $table->integer('NT');
             $table->integer('L');
             $table->integer('A');
-            $table->integer('G');
-            $table->integer('O');
+            $table->double('G');
+            $table->double('O');
             $table->integer('V');
         });
     }
